Comments, Bug Fixes, and quality of life changes
Added comments for the card properties and card input scripts, fixed several bugs with the Taylor series and slope field card interactions with the set inspection function and the remove child function, and added sprite layer management which moves the parent cards and the cards the player currently owns to the top of the playing field. Removed unused components from card. Cleaned up the scripts folder and made it easier to tell which scripts the game uses. As well as several other small changes.