Live Demo Bug fixes
Fixed the bug where a set of cards would be removed from the full set corral with having the scale reset if a card was de-spawned using the buttons. 

Added an offset to the position of the child cards of a set so the player can see how many cards are in a set without needing to inspect it.

updated Ts.Blue and Sf.Blue cards to add offsets.

updated card_properties documentation to reflect the updates.

Created a new build named 6-10 post-presentation build and placed it in the build folder.
Note: The style and index files will need to be updated before the new build can be placed on itch.

Created a new folder called 6-9 live demo build in the build folder and moved the existing build files for the current live build on itch into it.